Player AI in own class some minor changes GameLevel and KwdFile class changed KwdHeader class
